chroe: dir arrangement
