+++ assets/css/style.css
@@ -161,7 +161,7 @@ main {
 }
 
 .content3 .img-box {
-    background-image: linear-gradient(rgba(0,0,0,0.5) , rgba(0,0,0,0.5) ), url(./images/middle2-bg.png);
+    background-image: linear-gradient(rgba(0,0,0,0.5) , rgba(0,0,0,0.5) ), url(./assets/images/middle2-bg.png);
     background-repeat: no-repeat;
     background-position: center;
     overflow: hidden;
