+++ assets/js/test.js
@@ -1,0 +1,4 @@
+const emailTest = 'test@codeit.com'
+const passwordTest = 'codeit101'
+
+export {emailTest, passwordTest}
