+++ index.html
@@ -12,17 +12,17 @@
     <meta name="twitter:title" content="Linkbrary" /> 
     <meta name="twitter:description" content="세상의 모든 정보를 쉽게 저장하고 관리해 보세요" /> 
     <meta name="twitter:image" content="https://media.makeameme.org/created/html-css-and.jpg " />
-    <link rel="stylesheet" href="reset.css">
-    <link rel="stylesheet" href="common.css">
-    <link rel="stylesheet" href="style.css">
+    <link rel="stylesheet" href="/assets/css/reset.css">
+    <link rel="stylesheet" href="/assets/css/common.css">
+    <link rel="stylesheet" href="/assets/css/style.css">
     <link rel="stylesheet" as="style" crossorigin href="https://cdn.jsdelivr.net/gh/orioncactus/pretendard@v1.3.8/dist/web/static/pretendard.css" />
     <title>Linkbrary</title>
 </head>
 <body>
     <header>
         <nav>
-            <a href="index.html"><img src="/images/icon/logo.png" width="133px" height="24px"></a>
-            <a href="signin.html" class="link">
+            <a href="index.html"><img src="/assets/images/icon/logo.png" width="133px" height="24px"></a>
+            <a href="/pages/signin/signin.html" class="link">
                 로그인
             </a>
         </nav>
@@ -34,10 +34,10 @@
                     <span>세상의 모든 정보</span>를 <br>
                     쉽게 저장하고 관리해 보세요
                 </div>
-                <a href="signup.html" class="link">
+                <a href="/pages/signin/signin.html" class="link">
                     링크 추가하기
                 </a>
-                <img src="/images/top.png" class="img">
+                <img src="/assets/images/top.png" class="img">
             </section>
         </section>
         <section class="bottom-content">
@@ -57,20 +57,20 @@
                         </p>
                     </div>
                     <div class="img-box">
-                        <img src="/images/Card1.png">
+                        <img src="/assets/images/Card1.png">
                         <div class="card-2-box">
-                            <img src="/images/Card2.png">
-                            <img src="/images/icon/Cardicon.png" class="card2-icon">
+                            <img src="/assets/images/Card2.png">
+                            <img src="/assets/images/icon/Cardicon.png" class="card2-icon">
                             <div class="card2-border"></div>
                         </div>
-                        <img src="/images/Card3.png" class="card-3">
+                        <img src="/assets/images/Card3.png" class="card-3">
                     </div>
                 </div>
             </section>
             <section class="content2">
                 <div class="box">
                     <div class="img-box">
-                        <img src="/images/middle1.png" class="img">
+                        <img src="/assets/images/middle1.png" class="img">
                     </div>
                     <div class="text">
                         <p>
@@ -99,14 +99,14 @@
                         </p>
                     </div>
                     <div class="img-box">
-                        <img src="/images/middle2.png" class="img">
+                        <img src="/assets/images/middle2.png" class="img">
                     </div>
                 </div>
             </section>
             <section class="content4">
                 <div class="box">
                     <div class="img-box">
-                        <img src="/images/middle3.png" class="img">
+                        <img src="/assets/images/middle3.png" class="img">
                     </div>
                     <div class="text">
                         <p>
@@ -141,10 +141,10 @@
                 </a>
             </div>
             <div class="sns">
-                <a href="https://www.facebook.com/" target="_blank"><img src="/images//icon/akar-icons_facebook-fill.png"></a>
-                <a href="https://twitter.com/" target="_blank"><img src="/images//icon/akar-icons_twitter-fill.png"></a>
-                <a href="https://www.youtube.com/" target="_blank"><img src="/images//icon/akar-icons_youtube-fill.png"></a>
-                <a href="https://www.instagram.com/" target="_blank"><img src="/images//icon/ant-design_instagram-filled.png"></a>
+                <a href="https://www.facebook.com/" target="_blank"><img src="/assets/images//icon/facebook-fill.png"></a>
+                <a href="https://twitter.com/" target="_blank"><img src="/assets/images//icon/twitter-fill.png"></a>
+                <a href="https://www.youtube.com/" target="_blank"><img src="/assets/images//icon/youtube-fill.png"></a>
+                <a href="https://www.instagram.com/" target="_blank"><img src="/assets/images//icon/instagram-filled.png"></a>
             </div>
         </div>
     </footer>
