+++ pages/signin/signin.html
@@ -3,9 +3,9 @@
 <head>
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <link rel="stylesheet" href="reset.css">
-    <link rel="stylesheet" href="common.css">
-    <link rel="stylesheet" href="sign.css">
+    <link rel="stylesheet" href="/assets/css/reset.css">
+    <link rel="stylesheet" href="/assets/css/common.css">
+    <link rel="stylesheet" href="/assets/css/sign.css">
     <link rel="stylesheet" as="style" crossorigin href="https://cdn.jsdelivr.net/gh/orioncactus/pretendard@v1.3.8/dist/web/static/pretendard.css" />
     <title>signin</title>
 </head>
@@ -13,16 +13,16 @@
     <main class="sign" id="in">
         <section class="content">
             <div class="title">
-                <img src="/images/icon/signin-logo.png">
+                <img src="/assets/images/icon/signin-logo.png">
                 <div class="text">
-                    회원이 아니신가요? <a href="siginup.html">회원 가입하기</a>
+                    회원이 아니신가요? <a href="/pages/signup/siginup.html">회원 가입하기</a>
                 </div>
             </div>
             <div class="input">
                 <div class="box">
                     이메일
                     <input id="email" type="email" autocomplete="on" class="normal">
-                    <div class="error-text">
+                    <div class="error-text" id="email-error">
                         이메일을 입력해주세요.
                     </div>
                 </div>
@@ -30,9 +30,9 @@
                     비밀번호
                     <div class="input-box">
                         <input id="password" type="password" autocomplete="on" class="normal">
-                        <img src="/images/icon/signin-eye-off.png" class="icon-eye off">
+                        <img src="/assets/images/icon/signin-eye-off.png" class="icon-eye off">
                     </div>
-                    <div class="error-text">
+                    <div class="error-text" id="password-error">
                         비밀번호를 입력해주세요.
                     </div>
                 </div>
@@ -50,19 +50,19 @@
             <div class="logo">
                 <a href="https://www.google.com/">
                     <div class="box">
-                        <img src="/images/icon/signin-google-bg.png">
-                        <img src="/images/icon/signin-google.png" class="in-logo">
+                        <img src="/assets/images/icon/signin-google-bg.png">
+                        <img src="/assets/images/icon/signin-google.png" class="in-logo">
                     </div>
                 </a>
                 <a href="https://www.kakaocorp.com/page/">
                     <div class="box">
-                        <img src="/images/icon/signin-kakao.bg.png">
-                        <img src="/images/icon/signin-kakao.png" class="in-logo">
+                        <img src="/assets/images/icon/signin-kakao.bg.png">
+                        <img src="/assets/images/icon/signin-kakao.png" class="in-logo">
                     </div>
                 </a>
             </div>
         </section>
     </main> 
-    <script src="/sign.js"></script>
+    <script type="module" src="/pages/signin/signin.js"></script>
 </body>
 </html>
