--- siginup.html
@@ -1,71 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <link rel="stylesheet" href="reset.css">
-    <link rel="stylesheet" href="common.css">
-    <link rel="stylesheet" href="sign.css">
-    <link rel="stylesheet" as="style" crossorigin href="https://cdn.jsdelivr.net/gh/orioncactus/pretendard@v1.3.8/dist/web/static/pretendard.css" />
-    <title>signup</title>
-</head>
-<body>
-    <main class="sign" id="up">
-        <section class="content">
-            <div class="title">
-                <img src="/images/icon/signin-logo.png">
-                <div class="text">
-                    이미 회원이신가요? <a href="signin.html">로그인 하기</a>
-                </div>
-            </div>
-            <div class="input">
-                <div class="box">
-                    이메일
-                    <input id="email" type="email" autocomplete="on">
-                </div>
-                <div class="box">
-                    비밀번호
-                    <div class="input-box">
-                        <input id="password" type="password" autocomplete="on">
-                        <img src="/images/icon/signin-eye-off.png">
-                    </div>
-                </div>
-                <div class="box">
-                    비밀번호 확인
-                    <div class="check">
-                        <div class="input-box error">
-                            <input id="password" type="password" autocomplete="on">
-                            <img src="/images/icon/signin-eye-off.png">
-                        </div>
-                        비밀번호가 다릅니다
-                    </div>
-                </div>
-            </div> 
-            <a href="\">
-                <div class="login">
-                    회원가입 
-                </div>
-            </a>
-        </section>
-        <section class="sns">
-            <div class="title">
-                다른 방식으로 가입하기
-            </div>
-            <div class="logo">
-                <a href="https://www.google.com/">
-                    <div class="box">
-                        <img src="/images/icon/signin-google-bg.png">
-                        <img src="/images/icon/signin-google.png" class="in-logo">
-                    </div>
-                </a>
-                <a href="https://www.kakaocorp.com/page/">
-                    <div class="box">
-                        <img src="/images/icon/signin-kakao.bg.png">
-                        <img src="/images/icon/signin-kakao.png" class="in-logo">
-                    </div>
-                </a>
-            </div>
-        </section>
-    </main> 
-</body>
-</html>
